requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
joblib==1.1.0
nltk==3.7
numpy==1.22.4
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.4.24
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
